+++ Sources/MapboxNavigation/NavigationMapView.swift
@@ -56,6 +56,14 @@ open class NavigationMapView: UIView {
             }
         }
     }
+
+    /**
+     The tolerance value used for configuring the underlying map source of route line, maneuver arrow and restricted areas.
+
+     Controls the level of simplification by specifying the maximum allowed distance between the original line point and the simplified point. A higher tolerance value results in higher simplification and faster performance.
+     Changing the property will affect only newly added sources.
+     */
+    public var overlaySimplificationTolerance: Double = 0.375
     
     /**
      Controls whether to show fading gradient color on route lines between two different congestion
@@ -346,6 +354,7 @@ open class NavigationMapView: UIView {
                 
                 var arrowSource = GeoJSONSource()
                 arrowSource.data = .feature(Feature(geometry: .lineString(shaftPolyline)))
+                arrowSource.tolerance = overlaySimplificationTolerance
                 var arrowLayer = LineLayer(id: NavigationMapView.LayerIdentifier.arrowLayer)
                 if mapView.mapboxMap.style.sourceExists(withId: NavigationMapView.SourceIdentifier.arrowSource) {
                     let geoJSON = Feature(geometry: .lineString(shaftPolyline))
@@ -371,6 +380,7 @@ open class NavigationMapView: UIView {
                 
                 var arrowStrokeSource = GeoJSONSource()
                 arrowStrokeSource.data = .feature(Feature(geometry: .lineString(shaftPolyline)))
+                arrowStrokeSource.tolerance = overlaySimplificationTolerance
                 var arrowStrokeLayer = LineLayer(id: NavigationMapView.LayerIdentifier.arrowStrokeLayer)
                 if mapView.mapboxMap.style.sourceExists(withId: NavigationMapView.SourceIdentifier.arrowStrokeSource) {
                     let geoJSON = Feature(geometry: .lineString(shaftPolyline))
@@ -393,6 +403,7 @@ open class NavigationMapView: UIView {
                 let point = Point(shaftStrokeCoordinates.last!)
                 var arrowSymbolSource = GeoJSONSource()
                 arrowSymbolSource.data = .feature(Feature(geometry: .point(point)))
+                arrowSymbolSource.tolerance = overlaySimplificationTolerance
                 if mapView.mapboxMap.style.sourceExists(withId: NavigationMapView.SourceIdentifier.arrowSymbolSource) {
                     let geoJSON = Feature.init(geometry: Geometry.point(point))
                     try mapView.mapboxMap.style.updateGeoJSONSource(withId: NavigationMapView.SourceIdentifier.arrowSymbolSource,
@@ -533,6 +544,7 @@ open class NavigationMapView: UIView {
                 var restrictedAreaGeoJSON = GeoJSONSource()
                 restrictedAreaGeoJSON.data = .geometry(.lineString(shape))
                 restrictedAreaGeoJSON.lineMetrics = true
+                restrictedAreaGeoJSON.tolerance = overlaySimplificationTolerance
                 
                 try mapView.mapboxMap.style.addSource(restrictedAreaGeoJSON, id: sourceIdentifier)
             }
@@ -598,7 +610,7 @@ open class NavigationMapView: UIView {
         
         let geoJSONSource = self.geoJSONSource(shape)
         let sourceIdentifier = route.identifier(.source(isMainRoute: isMainRoute, isSourceCasing: true))
-        
+
         do {
             if mapView.mapboxMap.style.sourceExists(withId: sourceIdentifier) {
                 try mapView.mapboxMap.style.updateGeoJSONSource(withId: sourceIdentifier,
@@ -746,7 +758,7 @@ open class NavigationMapView: UIView {
         var geoJSONSource = GeoJSONSource()
         geoJSONSource.data = .geometry(.lineString(shape))
         geoJSONSource.lineMetrics = true
-        
+        geoJSONSource.tolerance = overlaySimplificationTolerance
         return geoJSONSource
     }